rename ACP to ACPInstance and move example_helpers.py to extras
