+++ examples/003_basic_rules.py
@@ -45,7 +45,7 @@ import tempfile
 # %%
 # Import the PyACP dependencies.
 from ansys.acp.core import ACPWorkflow, LinkedSelectionRule, launch_acp
-from ansys.acp.core.example_helpers import ExampleKeys, get_example_file
+from ansys.acp.core.extras.example_helpers import ExampleKeys, get_example_file
 
 # sphinx_gallery_thumbnail_number = -1
