+++ examples/004_advanced_rules.py
@@ -58,7 +58,7 @@ from ansys.acp.core import (
     example_helpers,
     launch_acp,
 )
-from ansys.acp.core.example_helpers import ExampleKeys, get_example_file
+from ansys.acp.core.extras.example_helpers import ExampleKeys, get_example_file
 
 # sphinx_gallery_thumbnail_number = 5
