+++ examples/006_rosettes.py
@@ -52,7 +52,7 @@ from ansys.acp.core import (
     get_directions_plotter,
     launch_acp,
 )
-from ansys.acp.core.example_helpers import ExampleKeys, get_example_file
+from ansys.acp.core.extras.example_helpers import ExampleKeys, get_example_file
 
 # sphinx_gallery_thumbnail_number = 4
