+++ src/ansys/acp/core/_server/__init__.py
@@ -20,7 +20,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from .acp_instance import ACP
+from .acp_instance import ACPInstance
 from .common import LaunchMode
 from .connect import ConnectLaunchConfig
 from .direct import DirectLaunchConfig
@@ -28,7 +28,7 @@ from .docker_compose import DockerComposeLaunchConfig
 from .launch import launch_acp
 
 __all__ = [
-    "ACP",
+    "ACPInstance",
     "ConnectLaunchConfig",
     "DirectLaunchConfig",
     "DockerComposeLaunchConfig",
