+++ src/ansys/acp/core/extras/__init__.py
@@ -1,0 +1,6 @@
+from ansys.acp.core.extras.example_helpers import get_example_file, ExampleKeys
+
+__all__ = [
+    "ExampleKeys",
+    "get_example_file",
+]
