+++ src/ansys/acp/core/extras/example_helpers.py
